Move argument parser options to settings module
This will help regroup them

Simplify building parser in main module

Adjust docstrings of option parsers

Add category for configuration options

Move deps-parser-choice to paths options

Move non-settings related options backj to main

Inspired by @jherland's comments

Add comment to why help message is generated separately

as sugested by @jherland in the review
